requirements.txt
glcontext==3.0.0
moderngl==5.12.0
numpy==2.3.5
pillow==12.0.0
pygame==2.6.1
pyglm==2.8.3
